+++ packages/stream_feed/test/faye_client_test.dart
@@ -1,6 +1,7 @@
 import 'dart:io';
 
-import 'package:stream_feed/src/client/stream_client.dart';
+import 'package:stream_feed/src/client/stream_feed_client.dart';
+import 'package:stream_feed/src/client/stream_feed_client_impl.dart';
 import 'package:stream_feed/src/core/models/feed_id.dart';
 import 'package:stream_feed/src/core/models/realtime_message.dart';
 import 'package:stream_feed/stream_feed.dart';
@@ -13,7 +14,7 @@ void main() async {
     final appId = env['appId'];
     final apiKey = env['apiKey'];
 
-    final client = StreamClient.connect(
+    final client = StreamFeedClient.connect(
       apiKey!,
       secret: secret,
       appId: appId,