+++ packages/stream_feed/CHANGELOG.md
@@ -1,3 +1,12 @@
-## 0.0.1 TODO: Add release date.
+## 0.1.3: 17/05/2021
+
+- fix: EnrichedActivity Not Returning Reactions 
+- update links in readme
+
+## 0.1.2: 07/05/2021
+
+- update dependencies
+
+## 0.1.1: 07/05/2021
 
 - first beta version